requirements.txt
sanic
sanic-ext
sanic-aiomysql
orjson
ipcs
tempylate
discord
async-stripe
toml